Support foreign keys introspection
This PR implements the following methods on the `DatabaseMetadata`:

 - `getImportedKeys`
 - `getExportedKeys`
 - `getCrossReference`

Testing: new tests added for foreign keys methods.

Fixes: #149
